+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Fix SSR and consider that selectedItem might be undefined.
+- Add fallback values to `measurementUnit` and `unitMultiplier`.
 
 ## [1.4.0] - 2020-04-28