+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.1] - 2020-05-01
 ### Fixed
 - Fix SSR and consider that selectedItem might be undefined.
 - Add fallback values to `measurementUnit` and `unitMultiplier`.